requirements.txt
azure-storage-blob==12.19.1 
Flask==3.0.3
keras==3.1.1
muspy==0.5.0
numpy==1.26.4
tensorflow-cpu==2.16.1 
gunicorn==22.0.0
pypianoroll